requirements.txt
# StarStitch Dependencies
# Core
python-dotenv>=1.0.0
requests>=2.31.0

# AI Providers
replicate>=0.25.0
fal-client>=0.4.0

# Video Processing
# Note: FFMPEG must be installed separately on your system

# Web UI (v0.2) - Streamlit
streamlit>=1.31.0
streamlit-sortables>=0.2.0
watchdog>=3.0.0

# API Server (v0.6) - FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
pydantic-settings>=2.1.0
aiofiles>=23.2.1
